タブを３区分化・カードを並べ替え・「更新を終了したグラフ」の折りたたみを実装
