+++ components/index/CardsFeatured/ConfirmedCasesDetails/Table.vue
