+++ components/index/CardsFeatured/ConfirmedCasesNumber/Card.vue
