+++ components/index/CardsFeatured/TestedNumber/Card.vue
@@ -51,7 +51,7 @@
 import dayjs from 'dayjs'
 import duration from 'dayjs/plugin/duration'
 
-import Chart from '@/components/index/CardsReference/TestedNumber/Chart.vue'
+import Chart from '@/components/index/CardsFeatured/TestedNumber/Chart.vue'
 import Data from '@/data/data.json'
 dayjs.extend(duration)
